feat: Refactor Spotify card generation and add new endpoint for default layout
