+++ vercel.json
@@ -1,19 +1,23 @@
 {
-    "version": 2,
-    "builds": [
-        {
-            "src": "src/index.js",
-            "use": "@vercel/node"
-        }
-    ],
-    "routes": [
-        {
-            "src": "/api",
-            "dest": "src/index.js"
-        },
-        {
-            "src": "/(.*)",
-            "dest": "src/index.js"
-        }
-    ]
+  "version": 2,
+  "builds": [
+    {
+      "src": "src/index.js",
+      "use": "@vercel/node"
+    }
+  ],
+  "routes": [
+    {
+      "src": "/api",
+      "dest": "src/index.js"
+    },
+    {
+      "src": "/card(.*)",
+      "dest": "src/index.js"
+    },
+    {
+      "src": "/(.*)",
+      "dest": "src/index.js"
+    }
+  ]
 }
